+++ mycollab-web/src/main/java/com/esofthead/mycollab/module/project/view/bug/components/UnresolvedBugsByAssigneeWidget.java
@@ -77,7 +77,8 @@ public class UnresolvedBugsByAssigneeWidget extends DepotWithChart {
                 }
                 BugAssigneeLink userLbl = new BugAssigneeLink(assignUser, item.getExtraValue(), assignUserFullName);
                 assigneeLayout.addComponent(userLbl);
-                ProgressBarIndicator indicator = new ProgressBarIndicator(totalCount, totalCount - item.getValue(), false);
+                ProgressBarIndicator indicator = new ProgressBarIndicator(totalCount, totalCount - item.getValue().intValue(),
+                        false);
                 indicator.setWidth("100%");
                 assigneeLayout.with(indicator).expand(indicator);
                 bodyContent.addComponent(assigneeLayout);
