+++ mycollab-web/src/main/java/com/esofthead/mycollab/module/project/view/task/components/UnresolvedTaskByAssigneeWidget.java
@@ -113,7 +113,8 @@ public class UnresolvedTaskByAssigneeWidget extends DepotWithChart {
 
                 TaskAssigneeLink taskAssigneeLink = new TaskAssigneeLink(assignUser, item.getExtraValue(), assignUserFullName);
                 assigneeLayout.addComponent(new MCssLayout(taskAssigneeLink).withWidth("110px"));
-                ProgressBarIndicator indicator = new ProgressBarIndicator(totalCountItems, totalCountItems - item.getValue(), false);
+                ProgressBarIndicator indicator = new ProgressBarIndicator(totalCountItems, totalCountItems - item.getValue().intValue(),
+                        false);
                 indicator.setWidth("100%");
                 assigneeLayout.with(indicator).expand(indicator);
                 bodyContent.addComponent(assigneeLayout);
